tests: update accessibility tests
Tests updated according to upstream code
